loops.py commits

+++ AskUserForInput.py
@@ -2,8 +2,8 @@ salary=input("Kindly enter your salary amount ")
 bonus=input("Kindly enter your desired bonus amount ")
 #payCheckAmount=float (salary)+float(bonus)
 #(2*float(bonus))
-s=float(salary)
-b=float(bonus)
+s=int(salary)
+b=int(bonus)
 payCheckAmount=0
 payCheckAmount=s+b
 print("The PayCheckAmount is {0:.2f} ".format(payCheckAmount))
